refactor(product): Product 가격 타입 Long에서 BigDecimal로 변경 - Product 엔티티 price 필드 BigDecimal 전환 및 precision/scale 설정 - ProductCreateRequest, ProductUpdateRequest, ProductResponse DTO 반영 - ProductFixture 및 Product 관련 테스트 코드 수정

+++ src/main/kotlin/com/hoppingmall/mall/product/dto/response/ProductWithImageResponse.kt
@@ -3,6 +3,7 @@ package com.hoppingmall.mall.product.dto.response
 import com.hoppingmall.mall.global.enums.ProductStatus
 import com.hoppingmall.mall.product.domain.Product
 import com.hoppingmall.mall.product.domain.ProductImage
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class ProductResponse(
@@ -10,7 +11,7 @@ data class ProductResponse(
     val sellerId: Long,
     val name: String,
     val description: String,
-    val price: Long,
+    val price: BigDecimal,
     val status: ProductStatus,
     val imageUrl: String?,
     val createdAt: LocalDateTime,
